+++ tests/test_authentication.py
@@ -35,9 +35,8 @@ class AuthenticationTestCase(TestCase):
             with self.assertRaises(
                 pyicloud_ipd.exceptions.PyiCloudFailedLoginException
             ) as context:
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS)(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
                     "bad_username",
-                    "bad_password",
                     cookie_dir,
                     False,
                     "EC5646DE-9423-11E8-BF21-14109FE0B321",
@@ -59,9 +58,8 @@ class AuthenticationTestCase(TestCase):
                 # delete ./tests/vcr_cassettes/auth_requires_2sa.yml,
                 # put your actual credentials in here, run the test,
                 # and then replace with dummy credentials.
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS)(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
                     "jdoe@gmail.com",
-                    "password1",
                     cookie_dir,
                     True,
                     "DE309E26-942E-11E8-92F5-14109FE0B321",
@@ -85,9 +83,8 @@ class AuthenticationTestCase(TestCase):
                 # delete ./tests/vcr_cassettes/auth_requires_2fa.yml,
                 # put your actual credentials in here, run the test,
                 # and then replace with dummy credentials.
-                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS)(
+                authenticator(setup_logger(), "com", identity, lp_filename_concatinator, RawTreatmentPolicy.AS_IS, {})(
                     "jdoe@gmail.com",
-                    "password1",
                     cookie_dir,
                     True,
                     "EC5646DE-9423-11E8-BF21-14109FE0B321",